test: basic e2e for pull command
