+++ lib/consts.ts
@@ -1,20 +1,22 @@
 import { homedir } from "os";
 import path from "path";
 
-export const API_HOST =
-  process.env.DITTO_API_HOST || "https://api.dittowords.com";
-export const CONFIG_FILE =
-  process.env.DITTO_CONFIG_FILE || path.join(homedir(), ".config", "ditto");
-export const PROJECT_CONFIG_FILE = path.normalize(
-  path.join("ditto", "config.yml")
-);
-export const TEXT_DIR = process.env.DITTO_TEXT_DIR || "ditto";
-export const TEXT_FILE = path.normalize(path.join(TEXT_DIR, "text.json"));
-
-export default {
-  API_HOST,
-  CONFIG_FILE,
-  PROJECT_CONFIG_FILE,
-  TEXT_DIR,
-  TEXT_FILE,
-};
+export default new (class {
+  get API_HOST() {
+    return process.env.DITTO_API_HOST || "https://api.dittowords.com";
+  }
+  get CONFIG_FILE() {
+    return (
+      process.env.DITTO_CONFIG_FILE || path.join(homedir(), ".config", "ditto")
+    );
+  }
+  get PROJECT_CONFIG_FILE() {
+    return path.normalize(path.join("ditto", "config.yml"));
+  }
+  get TEXT_DIR() {
+    return process.env.DITTO_TEXT_DIR || "ditto";
+  }
+  get TEXT_FILE() {
+    return path.normalize(path.join(this.TEXT_DIR, "text.json"));
+  }
+})();
