+++ lib/http/__mocks__/fetchComponentFolders.ts
@@ -1,18 +1,23 @@
 import type { FetchComponentFoldersResponse } from "../fetchComponentFolders";
 
+export const MOCK_COMPONENT_FOLDERS: FetchComponentFoldersResponse = {
+  example_folder: "Example Folder",
+};
+
+export const MOCK_COMPONENT_FOLDERS_WITH_SAMPLE_DATA: FetchComponentFoldersResponse =
+  {
+    example_folder: "Example Folder",
+    example_folder_sample: "Sample Example Folder",
+  };
+
 export async function fetchComponentFolders(
   _options: {
     showSampleData?: boolean;
   } = {}
 ): Promise<FetchComponentFoldersResponse> {
   if (_options.showSampleData) {
-    return {
-      example_folder: "Example Folder",
-      example_folder_sample: "Sample Example Folder",
-    };
+    return MOCK_COMPONENT_FOLDERS_WITH_SAMPLE_DATA;
   }
 
-  return {
-    example_folder: "Example Folder",
-  };
+  return MOCK_COMPONENT_FOLDERS;
 }
