+++ lib/http/__mocks__/fetchVariants.ts
@@ -1,19 +1,21 @@
 import { IVariant } from "../fetchVariants";
 
+export const MOCK_VARIANTS_RESPONSE: IVariant[] = [
+  {
+    name: "Example Variant 1",
+    description: "This is example variant 1.",
+    apiID: "example-variant-1",
+  },
+  {
+    name: "Example Variant 2",
+    description: "This is example variant 2.",
+    apiID: "example-variant-2",
+  },
+];
+
 export async function fetchVariants(
-  source: any,
-  options: any = {}
+  _source: any,
+  _options: any = {}
 ): Promise<IVariant[] | null> {
-  return [
-    {
-      name: "Example Variant 1",
-      description: "This is example variant 1.",
-      apiID: "example-variant-1",
-    },
-    {
-      name: "Example Variant 2",
-      description: "This is example variant 2.",
-      apiID: "example-variant-2",
-    },
-  ];
+  return MOCK_VARIANTS_RESPONSE;
 }
