+++ lib/utils/getSelectedProjects.ts
@@ -1,7 +1,7 @@
 import fs from "fs";
 import yaml, { YAMLException } from "js-yaml";
 
-import { PROJECT_CONFIG_FILE } from "../consts";
+import consts from "../consts";
 import { ConfigYAML, Project } from "../types";
 import Config, { DEFAULT_CONFIG_JSON } from "../config";
 
@@ -29,8 +29,8 @@ function yamlToJson(_yaml: string): ConfigYAML | null {
  * Returns an array containing all valid projects ({ id, name })
  * currently contained in the project config file.
  */
-export const getSelectedProjects = (configFile = PROJECT_CONFIG_FILE) =>
+export const getSelectedProjects = (configFile = consts.PROJECT_CONFIG_FILE) =>
   Config.parseSourceInformation(configFile).validProjects;
 
-export const getIsUsingComponents = (configFile = PROJECT_CONFIG_FILE) =>
+export const getIsUsingComponents = (configFile = consts.PROJECT_CONFIG_FILE) =>
   Config.parseSourceInformation(configFile).shouldFetchComponentLibrary;
